Cleaning the rpmbuild tree to be able to skip.
deployment/docker/build.sh is getting the feature for cached build. They are copied to the container just before running the build script. But these lines remove the files from build cache which already exist.
"git archive" is used to copy the file tree so ".git/" does not exist in $OPENVNET_SRC_BUILD_DIR. The build process runs in the git repository where you cloned (/var/lib/jenkins/workspace/* for Jenkins case).

+++ deployment/packagebuild/build_packages_vnet.sh
@@ -43,25 +43,16 @@ sudo yum-builddep -y "$OPENVNET_SPEC_FILE"
 #
 
 OPENVNET_SRC_BUILD_DIR="${WORK_DIR}/SOURCES/openvnet"
-if [ -d "$OPENVNET_SRC_BUILD_DIR" ]; then
+if [[ -d "$OPENVNET_SRC_BUILD_DIR" && -z "${SKIP_CLEANUP}" ]]; then
   rm -rf "$OPENVNET_SRC_BUILD_DIR"
 fi
 
-mkdir -p "${WORK_DIR}/SOURCES"
-cp -r "$OPENVNET_SRC_ROOT_DIR" "${WORK_DIR}/SOURCES/openvnet"
-
-# Get rid up any possible dirty build directories
-for arch in "${POSSIBLE_ARCHS[@]}"; do
-  if [ -d "${WORK_DIR}/RPMS/${arch}" ]; then
-    rm -rf "${WORK_DIR}/RPMS/${arch}"
-  fi
-done
-
-
-# Clean up the source dir if it's dirty
-cd ${OPENVNET_SRC_BUILD_DIR}
-git reset --hard
-git clean -xdf
+mkdir -p "${OPENVNET_SRC_BUILD_DIR}"
+# Copy only the tracked files to rpmbuild SOURCES/.
+(
+  cd $(git rev-parse --show-toplevel)
+  git archive HEAD | tar x -C "${OPENVNET_SRC_BUILD_DIR}"
+)
 
 #
 # Build the packages
